Modified ScalapackWrapper class
